separated patient routes into routers file
